server/requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
certifi==2019.3.9
cffi==1.12.3
channels==2.2.0
channels-api==0.4.1
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-allauth==0.39.1
django-background-tasks==1.2.0
django-compat==1.0.15
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.9.4
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
oauthlib==3.0.1
psycopg2-binary==2.8.2
pycparser==2.19
PyHamcrest==1.9.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
urllib3==1.25.3
whitenoise==4.1.2
zope.interface==4.6.0
